pyproject.toml
[project]
name = "lsa-quamba-fork"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets==3.3.0",
    "flash-linear-attention==0.3.0",
    "holistictraceanalysis==0.2.0",
    "huggingface-hub==0.27.0",
    "numpy==1.26.4",
    "peft==0.10.0",
    "protobuf==4.25.2",
    "scikit-learn==1.6.1",
    "scipy==1.15.2",
    "sentencepiece==0.2.0",
    "torch==2.5.0",
    "torchao==0.12.0",
    "torchaudio==2.5.0",
    "torchvision==0.20.0",
    "transformers==4.45.0",
    "zstandard==0.23.0",
]
